Change int to float, file: dog.h

+++ 0x0E-structures_typedef/dog.h
@@ -1,7 +1,7 @@
 /**
 * struct dog - Short description
 * @name: char - Dog name
-* @age: int - Dog age
+* @age: float - Dog age
 * @owner: char - Dog owner
 *
 * Description: Struct of a Dog with three attributes 
@@ -9,6 +9,6 @@
 struct dog
 {
 char *name;
-int *age;
+float *age;
 char *owner;
 };
